server/checkout-manager-server/src/main/kotlin/com/wutsi/checkout/manager/util/NumberUtil.kt
package com.wutsi.checkout.manager.util

import java.text.CharacterIterator
import java.text.StringCharacterIterator

object NumberUtil {
    fun toHumanReadableByteCountSI(bytes: Long): String {
        var bytes = bytes
        if (-1000 < bytes && bytes < 1000) {
            return "$bytes B"
        }
        val ci: CharacterIterator = StringCharacterIterator("KMGTPE")
        while (bytes <= -999950 || bytes >= 999950) {
            bytes /= 1000
            ci.next()
        }
        return String.format("%.1f %cB", bytes / 1000.0, ci.current())
    }
}
